Add file search
