+++ components/Registry.tsx
@@ -20,6 +20,7 @@ import {
 import Header from "./Header";
 import Footer from "./Footer";
 import FileDisplay from "./FileDisplay";
+import DirectoryListing from "./DirectoryListing";
 import { CookieBanner } from "./CookieBanner";
 
 const Registry = () => {
@@ -312,12 +313,12 @@ const Registry = () => {
             } else {
               return (
                 <>
-                  {dirEntries && (
+                  {versionMeta && dirEntries && (
                     <DirectoryListing
                       name={name}
                       version={version}
                       path={path}
-                      dirEntries={dirEntries}
+                      dirListing={versionMeta.directoryListing}
                       repositoryURL={repositoryURL}
                     />
                   )}
@@ -500,150 +501,6 @@ function VersionSelector({
       ) : null}
     </div>
   );
-}
-
-function DirectoryListing(props: {
-  dirEntries: DirEntry[];
-  name: string;
-  version: string | undefined;
-  path: string;
-  repositoryURL?: string | null;
-}) {
-  const { asPath } = useRouter();
-  const isStd = asPath.startsWith("/std");
-  return (
-    <div className="flex flex-col pt-4">
-      <div className="-my-2 py-2 overflow-x-auto sm:-mx-6 sm:px-6 lg:-mx-8 lg:px-8">
-        <div className="inline-block min-w-full shadow-sm rounded-lg border border-gray-200 overflow-hidden">
-          <div className="bg-gray-100 border-b border-gray-200 py-2 px-4 flex justify-between">
-            <div className="flex items-center">
-              <svg
-                fill="currentColor"
-                viewBox="0 0 20 20"
-                className="w-6 h-6 text-gray-400 inline-block mr-2"
-              >
-                <path d="M2 6a2 2 0 012-2h5l2 2h5a2 2 0 012 2v6a2 2 0 01-2 2H4a2 2 0 01-2-2V6z"></path>
-              </svg>
-              <span className="ml-2 font-medium">{props.path || "/"}</span>
-            </div>
-            {props.repositoryURL && (
-              <a href={props.repositoryURL} className="link ml-4">
-                Repository
-              </a>
-            )}
-          </div>
-          <div>
-            <table className="min-w-full table-fixed">
-              <tbody className="bg-white">
-                {props.dirEntries
-                  .sort((a, b) => a.type.localeCompare(b.type))
-                  .map((entry, i) => {
-                    const href = `${isStd ? "" : "/x"}/[...rest]`;
-                    const as = encodeURI(
-                      `${isStd ? "" : "/x"}/${props.name}${
-                        props.version ? `@${props.version}` : ""
-                      }${props.path}/${entry.name}`
-                    );
-                    return (
-                      <tr
-                        key={i}
-                        className={`table-row hover:bg-gray-100${
-                          i !== props.dirEntries.length - 1
-                            ? " border-b border-gray-200"
-                            : ""
-                        }`}
-                      >
-                        <td className="whitespace-no-wrap text-sm leading-5 text-gray-400 w-6">
-                          <Link href={href} as={as}>
-                            <a
-                              className={`px-2 sm:pl-3 md:pl-4 py-1 w-full block ${
-                                entry.type === "dir"
-                                  ? "text-blue-300"
-                                  : "text-gray-300"
-                              }`}
-                              tabIndex={-1}
-                            >
-                              <svg
-                                fill="currentColor"
-                                viewBox="0 0 20 20"
-                                className="w-5 h-5"
-                              >
-                                {(() => {
-                                  switch (entry.type) {
-                                    case "file":
-                                      if (isReadme(entry.name)) {
-                                        return (
-                                          <path d="M9 4.804A7.968 7.968 0 005.5 4c-1.255 0-2.443.29-3.5.804v10A7.969 7.969 0 015.5 14c1.669 0 3.218.51 4.5 1.385A7.962 7.962 0 0114.5 14c1.255 0 2.443.29 3.5.804v-10A7.968 7.968 0 0014.5 4c-1.255 0-2.443.29-3.5.804V12a1 1 0 11-2 0V4.804z"></path>
-                                        );
-                                      }
-                                      return (
-                                        <path
-                                          fillRule="evenodd"
-                                          d="M4 4a2 2 0 012-2h4.586A2 2 0 0112 2.586L15.414 6A2 2 0 0116 7.414V16a2 2 0 01-2 2H6a2 2 0 01-2-2V4z"
-                                          clipRule="evenodd"
-                                        ></path>
-                                      );
-                                    case "dir":
-                                      return (
-                                        <path d="M2 6a2 2 0 012-2h5l2 2h5a2 2 0 012 2v6a2 2 0 01-2 2H4a2 2 0 01-2-2V6z"></path>
-                                      );
-                                    case "symlink":
-                                      return (
-                                        <path
-                                          fillRule="evenodd"
-                                          d="M12.586 4.586a2 2 0 112.828 2.828l-3 3a2 2 0 01-2.828 0 1 1 0 00-1.414 1.414 4 4 0 005.656 0l3-3a4 4 0 00-5.656-5.656l-1.5 1.5a1 1 0 101.414 1.414l1.5-1.5zm-5 5a2 2 0 012.828 0 1 1 0 101.414-1.414 4 4 0 00-5.656 0l-3 3a4 4 0 105.656 5.656l1.5-1.5a1 1 0 10-1.414-1.414l-1.5 1.5a2 2 0 11-2.828-2.828l3-3z"
-                                          clipRule="evenodd"
-                                        ></path>
-                                      );
-                                  }
-                                })()}
-                              </svg>
-                            </a>
-                          </Link>
-                        </td>
-                        <td
-                          className={
-                            "whitespace-no-wrap text-sm text-blue-500 leading-5" +
-                            (isReadme(entry.name) ? " font-medium" : "")
-                          }
-                        >
-                          <Link href={href} as={as}>
-                            <a className="px-2 py-1 w-full block">
-                              {entry.name}
-                            </a>
-                          </Link>
-                        </td>
-                        <td className="whitespace-no-wrap text-sm leading-5 text-gray-500 text-right">
-                          <Link href={href} as={as}>
-                            <a
-                              className="px-4 py-1 w-full h-full block"
-                              tabIndex={-1}
-                            >
-                              {entry.size ? (
-                                bytesToSize(entry.size)
-                              ) : (
-                                <>&nbsp;</>
-                              )}
-                            </a>
-                          </Link>
-                        </td>
-                      </tr>
-                    );
-                  })}
-              </tbody>
-            </table>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function bytesToSize(bytes: number) {
-  const sizes = ["B", "KB", "MB", "GB", "TB"];
-  if (bytes == 0) return "0 B";
-  const i = Math.floor(Math.log(bytes) / Math.log(1024));
-  return (bytes / Math.pow(1024, i)).toFixed(0) + " " + sizes[i];
 }
 
 export default Registry;
